перенес метод convert в отдельный модуль
